ex1-server-host-client/pyproject.toml
[project]
name = "ex1-server-host-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.14.1",
    "mcp>=1.25.0",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
]
